double changed to float

+++ include/cell_world/capture.h
@@ -11,19 +11,19 @@ namespace cell_world {
                 Add_member(angle);
                 Add_member(cool_down);
         );
-        double distance{};
-        double angle{};
-        double cool_down;
+        float distance{};
+        float angle{};
+        float cool_down;
     };
 
     struct Capture {
         Capture(const Capture_parameters &, const World &);
-        double dist_max;
-        double theta_diff_max;
+        float dist_max;
+        float theta_diff_max;
         Timer cool_down;
 
         Location_visibility visibility;
-        bool is_captured(const Location &predator_location, double predator_theta, const Location &prey_location);
+        bool is_captured(const Location &predator_location, float predator_theta, const Location &prey_location);
     };
 
     struct Cell_capture {
