+++ include/cell_world/visibility.h
@@ -11,11 +11,11 @@ namespace cell_world{
     };
 
     struct Coordinates_visibility_cone{
-        Coordinates_visibility_cone (const Graph &, double);
-        Cell_group visible_cells(const Cell &, double);
-        bool is_visible(const Cell &, double, const Cell &) const;
+        Coordinates_visibility_cone (const Graph &, float);
+        Cell_group visible_cells(const Cell &, float);
+        bool is_visible(const Cell &, float, const Cell &) const;
         const Graph &visibility;
-        const double visual_angle;
+        const float visual_angle;
     };
 
     struct Location_visibility{
